requirements.txt
langchain>=0.2.14
langchain-openai>=0.1.11
langchain-community>=0.2.12
langchain-ollama>=0.1.0
openai>=1.42.0
pydantic>=2.8.2
fastapi>=0.115.0
uvicorn>=0.30.6
httpx>=0.27.2
python-dotenv>=1.0.1
Pillow>=10.3.0
